trainer/consumer/requirements.txt
pandas==2.2.3
lightgbm==4.3.0
joblib==1.4.2
scikit-learn==1.4.2
kafka-python==2.0.2
requests
evidently==0.4.33
boto3==1.28.39
s3fs==2024.2.0
psutil==5.9.8
prometheus-client==0.20.0
redis==5.3.0
